+++ pkg/xz.sh
@@ -1,5 +1,5 @@
 [ -f Makefile ] && make clean
 CC="$MUSL_GCC" CFLAGS="$PKG_CFLAGS" LDFLAGS="$PKG_LDFLAGS" LIBS="$PKG_LIBS" STRIP="$STRIP" ./configure --host="$TARGET" --prefix=/usr --sysconfdir=/etc --localstatedir=/var && make install DESTDIR="$ROOT_DIR/bin/$ARCH/$PKG_NAME"
 STATUS=$?
-rm -fr "$ROOT_DIR/bin/$ARCH/$PKG_NAME/usr/share/doc/xz"
+[ $STATUS = $? ] && rm -fr "$ROOT_DIR/bin/$ARCH/$PKG_NAME/usr/share/doc/xz"
 [ $STATUS = $? ]
